Tidy up and small efficiency changes.
Tidied up the code a bit. Made an initialisation function for slices which can be used by both the Calculator class as well as children classes. Changed the calcuations to use the slice class variable, instead of passing an s variable between functions. Modified the pre calculation code to rehash( ) instead of reserve() , and moved the rehash( ) into the specialised doWork( ) functions ( the clear() remains in the preCalculate() function ).