Asynchronous slice initialisation
Changed the ThreadedCalculator class to call slice initialisation (initialiseSlice() function) asynchronously. The next slice initialisation is deployed asynchronously at the end of the computation of the previous slice, and is waited on at the beginning of the compuitation of the next slice. To properly accomodate this, the operator() function was changed to be virtual.

Also, generally cleaned up the rest of the code, removed #includes statements that were no longger necessary, as well as commented out code that is no longer needed.