Index Generator class
Added an index generator class, and a specialisation for it for the n=2 (D=2 in template parameters) case. This should work out (after a little tweaking) to be cleaner than the current method. In particular, the specialisation of n=2 allows us to get rid of the std::deque<idx_t> in favour of directly calculating the index from a given rank/position.

Doing so does not appear to have sped up the sequential calculator at all, but should improve the threaded calculator by removing the wait on the index generation (it is as yet unclear as to whether or not this is the case). It will use less memory.