Makefile improvement
Improved the makefile to use implicit rules where possible. This included renaming the .cc files to be <target>.cc. In the case of .so targets a new rule was written that uses the standard variables used by the implicit rules.

This allows more flexible adding of library paths or include paths when needed (such as when building on the carma and rcg servers at UoN).