Added virtual preCalculate()
Added a virtual preCalculate() function to the Calculator and ThreadedCalculator classes. This is a design improvement that better allows the ThreadedCalculator to make sure the parent class precalculation is performed before asynchronously calling the first initialseSlice() function. The virtual operator() was changed to a regular function in the Calculator class and was removed entirely from the ThreadedCalculator class.